# openapi.yaml
openapi: 3.1.0
info:
  title: Deklutter API
  description: AI-powered digital decluttering assistant. Clean Gmail inbox by automatically identifying and removing spam, newsletters, and promotional emails. More providers (Yahoo, Outlook, Drive, Dropbox) coming soon!
  version: 1.0.0
  contact:
    name: Deklutter Support
    email: mohammad.illiyaz@gmail.com

servers:
  - url: https://deklutter-api.onrender.com
    description: Production server

paths:
  /oauth/providers:
    get:
      operationId: listOAuthProviders
      summary: List supported OAuth providers
      description: Returns a list of all supported OAuth providers (Google, Yahoo, Dropbox, etc.)
      responses:
        '200':
          description: List of providers
          content:
            application/json:
              schema:
                type: object
                properties:
                  providers:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                          example: "google"
                        display_name:
                          type: string
                          example: "Google"
                        category:
                          type: string
                          example: "email"

  /oauth/{provider}/init:
    post:
      operationId: initOAuth
      summary: Initialize OAuth for any provider
      description: Start OAuth flow for Google, Yahoo, Dropbox, or any supported provider
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            example: "google"
          description: Provider name (google, yahoo, dropbox, etc.)
        - name: source
          in: query
          required: false
          schema:
            type: string
            default: "gpt"
            example: "gpt"
          description: Source identifier (gpt, web, mobile)
      responses:
        '200':
          description: OAuth URL generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  auth_url:
                    type: string
                    description: URL for user to authorize
                  provider:
                    type: string
                  source:
                    type: string

  /auth/google/init:
    post:
      operationId: initializeGmailOAuth
      summary: Initialize Gmail OAuth flow
      description: Generates an OAuth authorization URL for the user to grant Gmail access. Returns a URL that the user should visit to authorize. No authentication required - user identity established through OAuth.
      responses:
        '200':
          description: OAuth URL generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  auth_url:
                    type: string
                    description: URL for user to authorize Gmail access
                    example: "https://accounts.google.com/o/oauth2/auth?..."

  /gmail/scan:
    post:
      operationId: scanGmail
      summary: Scan Gmail inbox for unwanted emails
      description: Analyzes the user's Gmail inbox and classifies emails into categories (delete, review, keep). Returns statistics and email IDs for each category. User must have completed OAuth authorization first.
      security:
        - DeklutterOAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                days_back:
                  type: integer
                  description: Number of days to look back in inbox history
                  default: 30
                  minimum: 1
                  maximum: 365
                  example: 30
                limit:
                  type: integer
                  description: Maximum number of emails to scan
                  default: 100
                  minimum: 1
                  maximum: 1000
                  example: 100
      responses:
        '200':
          description: Scan completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: object
                    properties:
                      total_items:
                        type: integer
                        description: Total number of emails scanned
                        example: 127
                      counts:
                        type: object
                        properties:
                          delete:
                            type: integer
                            description: Number of emails safe to delete (spam/promotional)
                            example: 45
                          review:
                            type: integer
                            description: Number of emails that need review (uncertain)
                            example: 12
                          keep:
                            type: integer
                            description: Number of important emails to keep
                            example: 70
                      approx_size_mb:
                        type: number
                        format: float
                        description: Approximate size in MB that can be freed
                        example: 3.2
                  safe_to_delete:
                    type: array
                    description: Email IDs that are safe to delete
                    items:
                      type: string
                    example: ["msg_id_1", "msg_id_2"]
                  review:
                    type: array
                    description: Email IDs that need user review
                    items:
                      type: string
                    example: ["msg_id_3"]
                  keep:
                    type: array
                    description: Email IDs to keep
                    items:
                      type: string
                    example: ["msg_id_4", "msg_id_5"]
        '403':
          description: Gmail not authorized - User needs to complete OAuth flow first
        '500':
          description: Scan failed

  /gmail/apply:
    post:
      operationId: applyCleanup
      summary: Apply cleanup actions to Gmail
      description: Moves specified emails to trash or applies labels. Emails moved to trash can be recovered for 30 days. User must have completed OAuth authorization first.
      security:
        - DeklutterOAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message_ids
              properties:
                message_ids:
                  type: array
                  description: List of email IDs to process
                  items:
                    type: string
                  example: ["msg_id_1", "msg_id_2", "msg_id_3"]
                mode:
                  type: string
                  description: Action to perform on emails
                  enum: [trash, label_only]
                  default: trash
                  example: trash
      responses:
        '200':
          description: Cleanup applied successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  processed:
                    type: integer
                    description: Number of emails processed
                    example: 45
                  mode:
                    type: string
                    description: Action that was performed
                    example: trash
        '403':
          description: Gmail not authorized - User needs to complete OAuth flow first
        '500':
          description: Action failed

components:
  securitySchemes:
    DeklutterOAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://deklutter-api.onrender.com/auth/gpt/oauth/authorize
          tokenUrl: https://deklutter-api.onrender.com/auth/gpt/oauth/token
          scopes: {}
      description: OAuth 2.0 authorization for accessing user's Gmail through Deklutter